
\docType{data}
\name{camr}
\alias{camr}
\title{(Data) Under five and neonatal mortality in China by cause and age.}
\format{A data frame with 720 rows and 6 variables:
\describe{
  \item{agegp}{Age group of children}
  \item{reg}{Region of China (east, middle, or west)}
  \item{res}{rural or urban}
  \item{year}{Year of surveillance}
  \item{qc_nCH9}{Cummulation Mortality rate per live birth from birth up to oldest age due to pneumonia}
  \item{qc_nCH17}{Cummulation Mortality rate per live birth from birth up to oldest age due to injury}
  \item{qc_ntotal}{Cummulation Mortality rate per live birth from birth up to oldest age due to all causes}
    ...
}}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S2214109X16303345}
}
\usage{
u5mr
}
