% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombiningClassifier.R
\name{CombinedClassifier}
\alias{CombinedClassifier}
\title{Combined Classifier}
\usage{
CombinedClassifier(train_data, train_labels, test_data)
}
\arguments{
\item{train_data}{A data frame for training the classifiers.}

\item{train_labels}{A factor vector of labels for the training data.}

\item{test_data}{A data frame for testing the classifiers.}
}
\value{
A vector of predicted labels based on majority voting.
}
\description{
Combines logistic regression and decision tree classifiers to predict classification labels.
}
