% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create Profile.R
\name{createProfile}
\alias{createProfile}
\title{Create a summarised experiment object from a list of GRanges and a .bigWig file.}
\usage{
createProfile(bigWig, testRangesList, nOfWindowsList, samplename = NULL)
}
\arguments{
\item{bigWig}{A file path to a .bigWig file}

\item{testRangesList}{A GRangesList object containing the gene ranges to calculate scores over}

\item{nOfWindowsList}{A numeric vector containing the window lengths for binning. Each length in the list
must correspond to the same position in the GRanges list.}

\item{samplename}{The name of your sample, used to group data in the plot.}
}
\description{
This function takes a bigWig, GRangesList, list of window lengths, and a sample name.
It returns a summarised experiment object that can then be used for plotting.
}
