\name{divideme}
\alias{divideme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Divide the first number by the second
}
\description{
This function allows the user to divide the numbers in parameters. The first the number is divided by the second one
}
\usage{
divideme(a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{
A number
}
  \item{b}{
A number
}
}
\details{
If \code{x} or \code{y} is not numeric, the function will not run and will return a warning message
}
\value{
Return the division's result of \code{x} by \code{y}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
divideme(1, 2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
