% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filter_by_strand}
\alias{filter_by_strand}
\title{Filter a GRanges object by strandedness of the ranges}
\usage{
filter_by_strand(bed, plus = FALSE, minus = FALSE, other = FALSE)
}
\arguments{
\item{bed}{The GRanges object that you would like to filter}

\item{plus}{TRUE/FALSE whether you want to return the plus strand ranges}

\item{minus}{TRUE/FALSE whether you want to return the minus strand ranges}

\item{other}{TRUE/FALSE whether you want to return ranges with no strandedness - marked '*'}
}
\description{
Returns a GRanges object with only the ranges of the strandedness you want
}
