% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filterForIntrons}
\alias{filterForIntrons}
\title{Filter GRanges Objects With Block Counts >1}
\usage{
filterForIntrons(bed)
}
\arguments{
\item{bed}{The GRanges object that you would like to filter}
}
\description{
This function takes a GRanges object and then returns a new GRanges object
where all the ranges have more than 1 block count. I.e. they contain introns.
}
