% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeature.R
\name{getFeature}
\alias{getFeature}
\title{Get a specific feature of interest}
\usage{
getFeature(
  object,
  start_feature = "TSS",
  start_flank = 0,
  start_exon = NULL,
  start_exon_boundary = NULL,
  start_direction = "up",
  end_feature = "TES",
  end_flank = 0,
  end_direction = "down",
  end_exon = NULL,
  end_exon_boundary = NULL,
  return = "all"
)
}
\arguments{
\item{object}{The GRanges object to define your ranges from.}

\item{start_feature}{The feature you would like to start around.}

\item{start_flank}{How many bp to set as your flank region around the start.}

\item{start_exon}{If your start feature is an exon/intron, set this number to whichever you'd like to start around.}

\item{start_exon_boundary}{The exon boundary you'd like to start from. Set to '3prime' or '5prime'}

\item{start_direction}{The direction from the start point to set your flank region start.}

\item{end_feature}{The feature you'd like to end at.}

\item{end_flank}{How many bp to set as your flank region around the end point.}

\item{end_direction}{The direction from the end point to set your flank region start.}

\item{end_exon}{If your end feature is an exon/intron, set this number to whichever you'd like to end around.}

\item{end_exon_boundary}{The exon boundary you'd like to end at. Set to '3prime' or '5prime'}

\item{return}{Set the strandedness of the ranges you'd like to return. This can be set to 'all', 'plus', or 'minus'.}
}
\description{
Input the strand, start and end point of the range, and specify the feature you want this around. Speficy upstream
and downstream as well.
}
