% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get Ranges.R
\name{getRanges}
\alias{getRanges}
\title{Get Specific Gene Ranges From Your GRanges Object}
\usage{
getRanges(
  object,
  plus = TRUE,
  minus = FALSE,
  upstream = 2000,
  downstream = 2000,
  exon1 = TRUE,
  intron1 = TRUE,
  exon2Tes = TRUE
)
}
\arguments{
\item{object}{The GRanges object that you would like to get ranges from}

\item{plus}{TRUE/FALSE to only return ranges with plus strandedness}

\item{minus}{TRUE/FALSE to only return ranges with minus strandedness}

\item{upstream}{The width of the upstream range you want to return. Enter NULL if you don't want the upstream range.}

\item{downstream}{The width of the downstream range you want to return. Enter NULL if you don't want the downstream range.}

\item{exon1}{TRUE/FALSE to return the TSS->Exon1 range in your list}

\item{intron1}{TRUE/FALSE to return the intron1 range in your list}

\item{exon2Tes}{TRUE/FALSE to return the exon2 -> TES range}
}
\description{
This function returns a list of GRanges objects corresponding to specific gene ranges. Included in that list
is the upstream range, TSS->Exon1 range, Intron1 range, Exon2->TES range and the downstream range. This GRanges list
can then be used by the regionPlot and plotRegion functions to plot these specific regions on a graph. IMPORTANT: You must
filter your GRanges object to only include intron-containing genes before using this function
}
