% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges From File.R
\name{grangesFromFile}
\alias{grangesFromFile}
\title{Load A High .bed or a .gtf file}
\usage{
grangesFromFile(filePath, gtf = FALSE, bed = FALSE)
}
\arguments{
\item{filePath}{A file path to the .bed or .gtf file}

\item{gtf}{TRUE/FALSE if the file is a .gtf}

\item{bed}{TRUE/FALSE if the file is a .bed}
}
\description{
This function takes the file path of a .bed or .gtf file and returns a
GRanges object of the information in the file
}
