% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiply_mv.R
\name{Multiply_mv}
\alias{Multiply_mv}
\title{Multiplication of two square matrices and a vector}
\usage{
Multiply_mv(A, B, x, vector_first)
}
\arguments{
\item{A:}{a square matrix}

\item{B:}{a square matrix}

\item{x:}{a vector}

\item{vector_first:}{a parameter which decides the order of multiplication. E.g. If vector_first = True, the function will compute the product of a matrix and a vector. Else, it will first compute the product of two matrices.}
}
\value{
a product
}
\description{
This function performs multiplication of two square matrices and a vector
}
\examples{
A, B: 3*3 matrices
x: a vector with length of 3
Multiply_mv(A,B,x, vector_first = T)
}
