% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLSreg_SVD.R
\name{OLSreg.SVD}
\alias{OLSreg.SVD}
\title{OLS estimates of the regression coefficients using a SVD decomposition}
\usage{
OLSreg.SVD(x, y)
}
\arguments{
\item{x:}{design matrix, each columns represents a predict variable}

\item{y:}{dependent variable}
}
\value{
OLS estimates of coefficients in the regression model
}
\description{
This function obtains OLS estimates of coefficients using a SVD decomposition
}
\examples{
x: a 10*3 matrix
y: a vertor with length of 10
OLSreg.SVD(x,y)
}
