% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMVN.R
\name{rMVN}
\alias{rMVN}
\title{Generate multivariate normal data}
\usage{
rMVN(n, miu, sigma)
}
\arguments{
\item{n:}{sample size}

\item{miu:}{the mean vector of multivariate normal distribution}

\item{sigma:}{the covariance matrix of multivariate distribution}
}
\value{
data that follows MVN, and each row represents an observation
}
\description{
This function generates data that follow multivariate normal distribution
}
\examples{
n: 5
m: c(2,1,2)
v: diag(c(1,2,3))
rMVN(n,m,v)
}
