\name{substracteme}
\alias{substracteme}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Substraction of the 1st number by the 2nd
}
\description{
The 2nd number \code{y} is substracted to the first number \code{x}
}
\usage{
substracteme(a, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{a}{
A number
}
  \item{b}{
A number
}
}
\details{
The two numbers have to be positive to use the function, and the result too. (Both \code{x} and \code{y} have to be numbers)
}
\value{
The result is the substraction of the 1st number \code{x} by the second one \code{y}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
substracteme(5, 2) #GOOD
substracteme(5, 8) #BAD
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
