% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile Plot.R
\name{plotProfile}
\alias{plotProfile}
\title{Function To Plot Your Profile Plot objects}
\usage{
plotProfile(object, group = NULL, groupBy = NULL)
}
\arguments{
\item{object}{Profile Plot object you created}

\item{group}{= A list of gene names you'd like to facet your plots with}

\item{goupBy}{= A character vector of the group names to display on your plot}
}
\description{
This function takes the Profile Plot objects you've created and enables you to plot them.
}
